build: build the library with Vite and publish proper exports
- replace tsup with Vite in library mode; declarations come from tsc,
  since TypeScript 7 ships no JS API for a bundler plugin to use
- add an exports map with import/require/react-native conditions. Without
  it Node ESM always resolved to the CJS build: arethetypeswrong reported
  `node16 (from ESM)` as CJS, and now reports ESM
- drop terser along with its mangle.properties setting. It saved zero
  bytes (280 vs 281 B gzipped) while silently renaming any new public
  property, and Vite deliberately keeps whitespace in es-format lib
  builds because stripping it removes pure annotations and breaks
  tree-shaking
- collapse src into a single module. The split files made tsc emit
  cross-referencing declarations, whose extensionless imports failed to
  resolve under node16 ESM — the same check now passes on all four
  resolution modes
- clean up tsconfig: drop the removed baseUrl and the jest types that
  were never installed, move to moduleResolution bundler, and cover the
  tests so they are type-checked too
