+++ tsconfig.json
@@ -1,19 +1,18 @@
 {
   "compilerOptions": {
-    "baseUrl": "./",
     "module": "esnext",
     "target": "ESNext",
     "jsx": "react-jsx",
-    "moduleResolution": "node",
+    "moduleResolution": "bundler",
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
     "noImplicitThis": true,
     "noImplicitAny": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "strict": true
+    "strict": true,
+    "skipLibCheck": true
   },
-  "include": ["src"],
-  "exclude": ["node_modules", "lib"],
-  "types": ["jest", "node"]
+  "include": ["src", "playwright", "vite.config.ts", "playwright.config.ts"],
+  "exclude": ["node_modules", "lib"]
 }
